+++ src/reduct/bucket.h
@@ -20,6 +20,8 @@ namespace reduct {
  */
 class IBucket {
  public:
+  virtual ~IBucket() = default;
+
   enum class QuotaType { kNone, kFifo };
 
   /**
@@ -84,8 +86,8 @@ class IBucket {
    * @param stop stop time point in the interval
    * @return the list or communication error
    */
-  [[nodiscard]] virtual Result<std::vector<RecordInfo>> List(std::string_view entry_name, Time start,
-                                                             Time stop) const noexcept = 0;
+  [[nodiscard, deprecated("Will be deleted in v1.0.0. Use IBucket::Query")]] virtual Result<std::vector<RecordInfo>>
+  List(std::string_view entry_name, Time start, Time stop) const noexcept = 0;
 
   /**
    * Write an object to the storage with timestamp
@@ -130,6 +132,42 @@ class IBucket {
   virtual Error Write(std::string_view entry_name, std::optional<Time> ts, size_t content_length,
                       WriteCallback callback) const noexcept = 0;
 
+  struct QueryOptions {
+    std::chrono::milliseconds ttl;
+  };
+
+  /**
+   * Record
+   */
+  struct Record {
+    Time timestamp;
+    size_t size;
+    bool last;
+
+    /**
+     * Function to receive data in chunks
+     */
+    std::function<Error(ReadCallback)> Read;
+  };
+
+  /**
+   * Callback for Bucket::Query
+   */
+  using NextRecordCallback = std::function<bool(Record&& record)>;
+
+  /**
+   * Query data for time interval
+   * @param entry_name
+   * @param start start time point ,if nullopt then from very beginning
+   * @param stop stop time point, if nullopt then until the last record
+   * @param options
+   * @param callback return  next record If you want to stop querying, return false
+   * @return
+   */
+  [[nodiscard]] virtual Error Query(
+      std::string_view entry_name, std::optional<Time> start = std::nullopt, std::optional<Time> stop = std::nullopt,
+      std::optional<QueryOptions> options = std::nullopt,
+      NextRecordCallback callback = [](auto) { return false; }) const noexcept = 0;
   /**
    * @brief Get settings by HTTP request
    * @return settings or HTTP error